api-iam/requirements.txt
PyMySQL
DBUtils
diskcache
cryptography
PyJWT
requests
APScheduler
Flask
pyyaml~=6.0.2
ruamel.yaml~=0.18.6
ldap3
alibabacloud_tea_openapi
alibabacloud_tea_util
alibabacloud_dysmsapi20170525~=3.1.0
alibabacloud_ecs20140526
alibabacloud_cms20190101
tencentcloud-sdk-python-common
tencentcloud-sdk-python-sms
redis
redis-py-cluster
ldif
oss2
pytz
python-dateutil
paramiko